+++ App/Workout/Models/OverallStatistics.cs
@@ -6,7 +6,7 @@ namespace SaveApp.App.Workout.Models
     {
         public int NumberOfWorkouts { get; set; }
         public List<int>? WorkoutDuration { get; set; }
-        public List<int>? Volume { get; set; }
+        public List<double>? Volume { get; set; }
         public List<int>? TotalSets { get; set; }
         public List<int>? TotalReps { get; set; }
         public List<double>? BodyWeight { get; set; }
