+++ App/Workout/Models/WorkoutDetailsCreateInput.cs
@@ -3,7 +3,7 @@ namespace SaveApp.App.Workout.Models
     public class WorkoutDetailsCreateInput
     {
         public String? Name { get; set; }
-        public int? BodyWeight { get; set; }
+        public double? BodyWeight { get; set; }
         public DateTime? Date { get; set; }
         public DateTime? StartTime { get; set; }
         public DateTime? EndTime { get; set; }
